Google test update

+++ common/common_struct.h
@@ -3,13 +3,13 @@
 
 typedef struct state_b{
     /* 0:Not configured 1:Configured*/
-    int configured        :1; 
+    uint8_t configured        :1; 
     /* The struct created by malloc 0:No 1:Yes*/
-    int is_malloc         :1;
+    uint8_t is_malloc         :1;
     /* The struct freed or not 0:No 1:Yes*/
-    int is_free           :1;
+    uint8_t is_free           :1;
     /* If the struct be operated by multithread 0:No 1:Yes*/
-    int is_multithread    :1;
+    uint8_t is_multithread    :1;
     int _reserved         :4;
 }BIT_T;
